# openapi.yaml
openapi: 3.1.0
info:
  title: Country City DB API
  description: A comprehensive API for querying countries, states, cities, regions, and subregions worldwide.
  version: 1.0.0
servers:
  - url: /
paths:
...
  /api/v1/regions:
    get:
      summary: List regions
      operationId: listRegions
      tags: [Regions]
      parameters:
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/name"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/no_page"
      responses:
        "200":
          description: Paginated list of regions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedRegions"
    post:
      summary: List regions (JSON body)
      operationId: listRegionsPost
      tags: [Regions]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryParams"
      responses:
        "200":
          description: Paginated list of regions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedRegions"
  /api/v1/regions/{id}:
    get:
      summary: Get region by ID
      operationId: getRegion
      tags: [Regions]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Region object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Region"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/regions/{id}/subregions:
    get:
      summary: List subregions in a region
      operationId: listRegionSubregions
      tags: [Regions]
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/name"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/no_page"
      responses:
        "200":
          description: Paginated list of subregions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSubregions"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      summary: List subregions in a region (JSON body)
      operationId: listRegionSubregionsPost
      tags: [Regions]
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryParams"
      responses:
        "200":
          description: Paginated list of subregions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedSubregions"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/regions/{id}/countries:
    get:
      summary: List countries in a region
      operationId: listRegionCountries
      tags: [Regions]
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/name"
        - $ref: "#/components/parameters/iso2"
        - $ref: "#/components/parameters/iso3"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/no_page"
      responses:
        "200":
          description: Paginated list of countries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCountries"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      summary: List countries in a region (JSON body)
      operationId: listRegionCountriesPost
      tags: [Regions]
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryParams"
      responses:
        "200":
          description: Paginated list of countries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCountries"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ── Subregions ───────────────────────────────────────────
...
  /api/v1/countries:
    get:
      summary: List countries
      operationId: listCountries
      tags: [Countries]
      parameters:
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/name"
        - $ref: "#/components/parameters/iso2"
        - $ref: "#/components/parameters/iso3"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/no_page"
      responses:
        "200":
          description: Paginated list of countries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCountries"
    post:
      summary: List countries (JSON body)
      operationId: listCountriesPost
      tags: [Countries]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryParams"
      responses:
        "200":
          description: Paginated list of countries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCountries"
  /api/v1/countries/{id}:
    get:
      summary: Get country by ID
      operationId: getCountry
      tags: [Countries]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Country object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Country"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/countries/iso2/{code}:
    get:
      summary: Get country by ISO2 code
      operationId: getCountryByISO2
      tags: [Countries]
      parameters:
        - $ref: "#/components/parameters/code"
      responses:
        "200":
          description: Country object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Country"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/countries/iso3/{code}:
    get:
      summary: Get country by ISO3 code
      operationId: getCountryByISO3
      tags: [Countries]
      parameters:
        - $ref: "#/components/parameters/code"
      responses:
        "200":
          description: Country object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Country"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/countries/name/{name}:
    get:
      summary: Get country by name
      operationId: getCountryByName
      tags: [Countries]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Country name
      responses:
        "200":
          description: Country object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Country"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/countries/{id}/states:
    get:
      summary: List states in a country
      operationId: listCountryStates
      tags: [Countries]
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/name"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/no_page"
      responses:
        "200":
          description: Paginated list of states
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedStates"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      summary: List states in a country (JSON body)
      operationId: listCountryStatesPost
      tags: [Countries]
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryParams"
      responses:
        "200":
          description: Paginated list of states
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedStates"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/countries/{id}/cities:
    get:
      summary: List cities in a country
      operationId: listCountryCities
      tags: [Countries]
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/name"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/no_page"
      responses:
        "200":
          description: Paginated list of cities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCities"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      summary: List cities in a country (JSON body)
      operationId: listCountryCitiesPost
      tags: [Countries]
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryParams"
      responses:
        "200":
          description: Paginated list of cities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCities"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ── States ───────────────────────────────────────────────
  /api/v1/states:
    get:
      summary: List states
      operationId: listStates
      tags: [States]
      parameters:
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/name"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/no_page"
      responses:
        "200":
          description: Paginated list of states
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedStates"
    post:
      summary: List states (JSON body)
      operationId: listStatesPost
      tags: [States]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryParams"
      responses:
        "200":
          description: Paginated list of states
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedStates"
  /api/v1/states/{id}:
    get:
      summary: Get state by ID
      operationId: getState
      tags: [States]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: State object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/State"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/states/{id}/cities:
    get:
      summary: List cities in a state
      operationId: listStateCities
      tags: [States]
      parameters:
        - $ref: "#/components/parameters/id"
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/name"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/no_page"
      responses:
        "200":
          description: Paginated list of cities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCities"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      summary: List cities in a state (JSON body)
      operationId: listStateCitiesPost
      tags: [States]
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryParams"
      responses:
        "200":
          description: Paginated list of cities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCities"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ── Cities ───────────────────────────────────────────────
  /api/v1/cities:
    get:
      summary: List cities
      operationId: listCities
      tags: [Cities]
      parameters:
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/name"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - $ref: "#/components/parameters/no_page"
      responses:
        "200":
          description: Paginated list of cities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCities"
    post:
      summary: List cities (JSON body)
      operationId: listCitiesPost
      tags: [Cities]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryParams"
      responses:
        "200":
          description: Paginated list of cities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCities"
  /api/v1/cities/{id}:
    get:
      summary: Get city by ID
      operationId: getCity
      tags: [Cities]
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: City object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/City"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Stats ────────────────────────────────────────────────
...
components:
  securitySchemes:
    internalKey:
      type: apiKey
      in: header
      name: X-Internal-Key
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Resource ID
    code:
      name: code
      in: path
      required: true
      schema:
        type: string
      description: ISO country code
    search:
      name: search
      in: query
      schema:
        type: string
      description: Full-text search query
    name:
      name: name
      in: query
      schema:
        type: string
      description: Filter by exact name
    iso2:
      name: iso2
      in: query
      schema:
        type: string
      description: Filter by ISO2 code
    iso3:
      name: iso3
      in: query
      schema:
        type: string
      description: Filter by ISO3 code
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
      description: Maximum number of results (max 100)
    offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
      description: Number of results to skip
    no_page:
      name: no_page
      in: query
      schema:
        type: boolean
        default: false
      description: If true, return all results without pagination
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    QueryParams:
      type: object
      properties:
        search:
          type: string
        name:
          type: string
        iso2:
          type: string
        iso3:
          type: string
        limit:
          type: integer
          default: 20
        offset:
          type: integer
          default: 0
        no_page:
          type: boolean
          default: false
    Region:
      type: object
      required: [id, name, updated_at, flag]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        translations:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        flag:
          type: integer
        wiki_data_id:
          type: string
    Subregion:
      type: object
      required: [id, name, region_id, updated_at, flag]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        translations:
          type: string
        region_id:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        flag:
          type: integer
        wiki_data_id:
          type: string
    Country:
      type: object
      required: [id, name, updated_at, flag]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        iso3:
          type: string
        numeric_code:
          type: string
        iso2:
          type: string
        phonecode:
          type: string
        capital:
          type: string
        currency:
          type: string
        currency_name:
          type: string
        currency_symbol:
          type: string
        tld:
          type: string
        native:
          type: string
        population:
          type: integer
          format: int64
        gdp:
          type: integer
          format: int64
        region:
          type: string
        region_id:
          type: integer
          format: int64
        subregion:
          type: string
        subregion_id:
          type: integer
          format: int64
        nationality:
          type: string
        area_sq_km:
          type: number
          format: double
        postal_code_format:
          type: string
        postal_code_regex:
          type: string
        timezones:
          type: string
        translations:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        emoji:
          type: string
        emoji_u:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        flag:
          type: integer
        wiki_data_id:
          type: string
    State:
      type: object
      required: [id, name, country_id, country_code, updated_at, flag]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        country_id:
          type: integer
          format: int64
        country_code:
          type: string
        fips_code:
          type: string
        iso2:
          type: string
        iso3166_2:
          type: string
        type:
          type: string
        level:
          type: integer
        parent_id:
          type: integer
          format: int64
        native:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        timezone:
          type: string
        translations:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        flag:
          type: integer
        wiki_data_id:
          type: string
        population:
          type: string
    City:
      type: object
      required: [id, name, state_id, state_code, country_id, country_code, latitude, longitude, created_at, updated_at, flag]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        state_id:
          type: integer
          format: int64
        state_code:
          type: string
        country_id:
          type: integer
          format: int64
        country_code:
          type: string
        type:
          type: string
        level:
          type: integer
        parent_id:
          type: integer
          format: int64
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        native:
          type: string
        population:
          type: integer
          format: int64
        timezone:
          type: string
        translations:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        flag:
          type: integer
        wiki_data_id:
          type: string
...
    PaginatedRegions:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Region"
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
    PaginatedSubregions:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Subregion"
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
    PaginatedCountries:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Country"
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
    PaginatedStates:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/State"
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
    PaginatedCities:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/City"
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer